requirements.txt
Flask==2.3.3
Pillow==10.4.0 --only-binary :all:
